Added more rockets, added game progression,
Limit the total number of invaders on the field to 7. When an invader is eliminated, spawn a new one, until the the total number of invaders on the field and the eliminated number of invaders equals the total number of invaders of the level.

Number of rockets now scale with the level.

Speed of invaders now scale with level.

Added different music for different levels.